projects/spotify_watcher/pyproject.toml
[project]
name = "spotify_watcher"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Graham Doerksen", email = "46031209+gdoerksen@users.noreply.github.com" }
]
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
]

[project.scripts]
spotify_watcher = "spotify_watcher.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"